==> 02_calculate_fixation_variables.Rmd <==
---
title: "02 - Calculate fixation variables for analysis"
author: "Janet Bang"
date: "August 11, 2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


This code calculates the visual attention variables used in analyses (except for contingent looking).

Variables include:
1. slide.metrics: variables provided by GazeTracker (stimulus presentation software)
2. value: value provided by GazeTracker
3. participant: participant id
4. video_phase: phase of video (i.e., base, label1 or label2, cue1 or cue2, center, test, test_point)
5. object: fopam, mimole, nalip, pagoune
6. trial_num: 1, 2, 3, 4
7. motion: static or function
8. aoi: area of interest: eye, arw, target, competitor
9. group: ASD or TYP
10. seq: counterbalancing order (1, 2, 3, 4, 5, 6, 7, 8)
11. block: 1 or 2
12. cueCond: gaze or arrow
13. point.target: 1 = pointed correctly at target, 2 = pointed at competitor



# Load libraries
```{r}
library(tidyverse)
library(openxlsx)
library(reshape2)

theme_set(theme_bw())
```

# Read in data
```{r}
eyetk_wl_clean <- read_csv("./data/eyetk_wl_clean.csv")
```


# Select only the columns with the fixation duration variable and transform variabls
```{r}
# filter only fixation duration and transform
fix_dur <- eyetk_wl_clean %>% 
  filter(slide.metrics == "totalFixDur") %>% 
  mutate(slide.metrics = factor(slide.metrics), 
         participant = factor(participant), 
         video_phase = factor(video_phase), 
         motion = factor(motion),
         cueCond = factor(cueCond))
         
# reorder levels
fix_dur <- fix_dur %>% 
  mutate(video_phase = fct_relevel(video_phase, "base", "label1", "cue1", "label2", "cue2", "center", "test", "testPoint"), 
         motion = fct_relevel(motion, "static", "function"), 
         cueCond = fct_relevel(cueCond, "eye", "arw"))

```


# Turn to wide format to calculate proportions
```{r}
fix_dur_wide <- dcast(fix_dur, participant + object + trial_num + 
                      motion + group + seq + block +  cueCond + point.target
                      ~ video_phase + slide.metrics + aoi, value.var = "value")

# check # of participants: 22 ASD, 41 TD
fix_dur_wide %>% 
  distinct(participant, group) %>% 
  group_by(group) %>% 
  count()

```


# Calculate the proportion fixation duration to each aoi/total fixation duration to the slide
```{r}
## for a single proportion NA means that there was no data during the trial from GT
## rowSums: sum of an empty set (i.e., NA, NA = 0, by definition)
## rowMeans: product of an empty set (i.e., NA, NA = NaN)

# proportion fixation duration - target
fix_dur_wide <- fix_dur_wide %>% 
  mutate(base_propFixDur_target = base_totalFixDur_target / base_totalFixDur_overall,
         label1_propFixDur_target = label1_totalFixDur_target / label1_totalFixDur_overall,
           cue1_propFixDur_target = cue1_totalFixDur_target / cue1_totalFixDur_overall,
           label2_propFixDur_target = label2_totalFixDur_target / label2_totalFixDur_overall,
           cue2_propFixDur_target = cue2_totalFixDur_target / cue2_totalFixDur_overall,
           center_propFixDur_target = center_totalFixDur_target / center_totalFixDur_overall,
           test_propFixDur_target = test_totalFixDur_target / test_totalFixDur_overall,
           testPoint_propFixDur_target = testPoint_totalFixDur_target / testPoint_totalFixDur_overall)

fix_dur_wide <- fix_dur_wide %>% 
  mutate(label1and2sum_totalFixDur_target = rowSums(fix_dur_wide[c("label1_totalFixDur_target", "label2_totalFixDur_target")], na.rm = T), 
         label1and2sum_totalFixDur_overall = rowSums(fix_dur_wide[c("label1_totalFixDur_overall", "label2_totalFixDur_overall")], na.rm = T), 
         label1and2sum_propFixDur_target = label1and2sum_totalFixDur_target/label1and2sum_totalFixDur_overall)

fix_dur_wide <- fix_dur_wide %>% 
  mutate(cue1and2sum_totalFixDur_target = rowSums(fix_dur_wide[c("cue1_totalFixDur_target", "cue2_totalFixDur_target")], na.rm = T), 
         cue1and2sum_totalFixDur_overall = rowSums(fix_dur_wide[c("cue1_totalFixDur_overall", "cue2_totalFixDur_overall")], na.rm = T), 
         cue1and2sum_propFixDur_target = cue1and2sum_totalFixDur_target/cue1and2sum_totalFixDur_overall)



# proportion fixation duration - competitor
fix_dur_wide <- fix_dur_wide %>% 
  mutate(base_propFixDur_competitor = base_totalFixDur_competitor / base_totalFixDur_overall,
         label1_propFixDur_competitor = label1_totalFixDur_competitor / label1_totalFixDur_overall,
           cue1_propFixDur_competitor = cue1_totalFixDur_competitor / cue1_totalFixDur_overall,
           label2_propFixDur_competitor = label2_totalFixDur_competitor / label2_totalFixDur_overall,
           cue2_propFixDur_competitor = cue2_totalFixDur_competitor / cue2_totalFixDur_overall,
           center_propFixDur_competitor = center_totalFixDur_competitor / center_totalFixDur_overall,
           test_propFixDur_competitor = test_totalFixDur_competitor / test_totalFixDur_overall,
           testPoint_propFixDur_competitor = testPoint_totalFixDur_competitor / testPoint_totalFixDur_overall)

fix_dur_wide <- fix_dur_wide %>% 
  mutate(label1and2sum_totalFixDur_competitor = rowSums(fix_dur_wide[c("label1_totalFixDur_competitor", "label2_totalFixDur_competitor")], na.rm = T), 
         label1and2sum_propFixDur_competitor = label1and2sum_totalFixDur_competitor/label1and2sum_totalFixDur_overall)

fix_dur_wide <- fix_dur_wide %>% 
  mutate(cue1and2sum_totalFixDur_competitor = rowSums(fix_dur_wide[c("cue1_totalFixDur_competitor", "cue2_totalFixDur_competitor")], na.rm = T), 
         cue1and2sum_propFixDur_competitor = cue1and2sum_totalFixDur_competitor/cue1and2sum_totalFixDur_overall)



# proportion fixation duration - eye
fix_dur_wide <- fix_dur_wide %>% 
  mutate(base_propFixDur_eye = base_totalFixDur_eye / base_totalFixDur_overall,
         label1_propFixDur_eye = label1_totalFixDur_eye / label1_totalFixDur_overall,
           cue1_propFixDur_eye = cue1_totalFixDur_eye / cue1_totalFixDur_overall,
           label2_propFixDur_eye = label2_totalFixDur_eye / label2_totalFixDur_overall,
           cue2_propFixDur_eye = cue2_totalFixDur_eye / cue2_totalFixDur_overall,
           center_propFixDur_eye = center_totalFixDur_eye / center_totalFixDur_overall)

fix_dur_wide <- fix_dur_wide %>% 
  mutate(label1and2sum_totalFixDur_eye = rowSums(fix_dur_wide[c("label1_totalFixDur_eye", "label2_totalFixDur_eye")], na.rm = T), 
         label1and2sum_propFixDur_eye = label1and2sum_totalFixDur_eye/label1and2sum_totalFixDur_overall)

fix_dur_wide <- fix_dur_wide %>% 
  mutate(cue1and2sum_totalFixDur_eye = rowSums(fix_dur_wide[c("cue1_totalFixDur_eye", "cue2_totalFixDur_eye")], na.rm = T), 
         cue1and2sum_propFixDur_eye = cue1and2sum_totalFixDur_eye/cue1and2sum_totalFixDur_overall)



# proportion fixation duration - arw
fix_dur_wide <- fix_dur_wide %>% 
  mutate(base_propFixDur_arw = base_totalFixDur_arw / base_totalFixDur_overall,
         label1_propFixDur_arw = label1_totalFixDur_arw / label1_totalFixDur_overall,
           cue1_propFixDur_arw = cue1_totalFixDur_arw / cue1_totalFixDur_overall,
           label2_propFixDur_arw = label2_totalFixDur_arw / label2_totalFixDur_overall,
           cue2_propFixDur_arw = cue2_totalFixDur_arw / cue2_totalFixDur_overall,
           center_propFixDur_arw = center_totalFixDur_arw / center_totalFixDur_overall)

fix_dur_wide <- fix_dur_wide %>% 
  mutate(label1and2sum_totalFixDur_arw = rowSums(fix_dur_wide[c("label1_totalFixDur_arw", "label2_totalFixDur_arw")], na.rm = T), 
         label1and2sum_propFixDur_arw = label1and2sum_totalFixDur_arw/label1and2sum_totalFixDur_overall)

fix_dur_wide <- fix_dur_wide %>% 
  mutate(cue1and2sum_totalFixDur_arw = rowSums(fix_dur_wide[c("cue1_totalFixDur_arw", "cue2_totalFixDur_arw")], na.rm = T), 
         cue1and2sum_propFixDur_arw = cue1and2sum_totalFixDur_arw/cue1and2sum_totalFixDur_overall)
```




# Calculate the target advantage score: target - comp / target + comp
```{r}
# NA = there is no data for the target or the competitor
# NaN for the targAdvtg = child did not look at target or competitor so can't get a number

fix_dur_wide <- fix_dur_wide %>% 
  mutate(base_targAdvtg = (base_totalFixDur_target - base_totalFixDur_competitor) / 
                          (base_totalFixDur_target + base_totalFixDur_competitor), 
         label1_targAdvtg = (label1_totalFixDur_target - label1_totalFixDur_competitor) / 
                            (label1_totalFixDur_target + label1_totalFixDur_competitor),
         cue1_targAdvtg = (cue1_totalFixDur_target - cue1_totalFixDur_competitor) / 
                          (cue1_totalFixDur_target + cue1_totalFixDur_competitor), 
         label2_targAdvtg = (label2_totalFixDur_target - label2_totalFixDur_competitor) / 
                            (label2_totalFixDur_target + label2_totalFixDur_competitor),
         cue2_targAdvtg = (cue2_totalFixDur_target - cue2_totalFixDur_competitor) / 
                          (cue2_totalFixDur_target + cue2_totalFixDur_competitor),
         center_targAdvtg = (center_totalFixDur_target - center_totalFixDur_competitor) / 
                            (center_totalFixDur_target + center_totalFixDur_competitor),
         test_targAdvtg = (test_totalFixDur_target - test_totalFixDur_competitor) / 
                          (test_totalFixDur_target + test_totalFixDur_competitor),
         testPoint_targAdvtg = (testPoint_totalFixDur_target - testPoint_totalFixDur_competitor) / 
                               (testPoint_totalFixDur_target + testPoint_totalFixDur_competitor), 
         label1and2sum_targAdvtg = (label1and2sum_totalFixDur_target - label1and2sum_totalFixDur_competitor)/
                                   (label1and2sum_totalFixDur_target + label1and2sum_totalFixDur_competitor), 
         cue1and2sum_targAdvtg = (cue1and2sum_totalFixDur_target - cue1and2sum_totalFixDur_competitor)/
                                 (cue1and2sum_totalFixDur_target + cue1and2sum_totalFixDur_competitor))

```



# Write to csv
```{r}
write_csv(fix_dur_wide, "./data/fix_dur_wide.csv")
```
