==> 07_analyses_latency.Rmd <==
---
title: "07 - Analyses for latency to test"
author: "Janet Bang"
date: "August 19, 2019"
output:
  html_document:
    toc: yes
    toc_float: yes
  pdf_document:
    toc: yes
  word_document:
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=TRUE, warning=FALSE, message=FALSE, cache=FALSE)
```


This code includes the analyses for latency values.


## Load libraries and functions
```{r}
library(tidyverse)
library(openxlsx)
library(lme4)
library(influence.ME)
library(emmeans)
library(lmerTest)
library(compute.es)
library(lattice)
library(pwr)
library(sjPlot)

theme_set(theme_bw())

source("./helper_functions/R_helperFunctions_WLdata.R")
```


# Latency
## Read in csv files
```{r}
eyetk_wl <- read_csv("./data/eyetk_wl.csv") 
demographics <- read.csv("./data/demographics.csv")
more_than_25pcttk <- read_csv("./data/more_than_25pcttk.csv")
exp_info_wl <- read_csv("./data/exp_info_wl.csv")
```


## Add language and exp info and select variables
```{r}
# select language variable
language <- demographics %>% 
  select(c(participant, language))

# select exp info variables and add language
exp_info_wl_merge <- exp_info_wl %>% 
  select(-c(practice, eyetk_inc)) %>% 
  mutate(participant = as.integer(participant), 
         trial_num = as.integer(trial_num)) %>% 
  left_join(language, by = "participant")

# select variables for latency
latency <- eyetk_wl %>% 
  filter(slide.metrics == "Duration before first fixation arrival (seconds)", 
         video_phase == "test", 
         aoi == "target") %>% 
  select(-slide.metrics) %>% 
  left_join(exp_info_wl_merge, by = c("participant", "object", "trial_num", "motion"))
```



## Keep phases where Percent Tracking > 25 % and remove short/long latencies
```{r}
latency_clean <- latency %>% 
  right_join(more_than_25pcttk, by = c("participant", "object", "motion", "trial_num")) %>% 
  filter(value > .200 & value < 4)
```



## Number of observations per participant
```{r}
latency_clean %>% 
  group_by(participant) %>% 
  dplyr::count()

# count per group
latency_clean %>% 
  distinct(participant, group) %>% 
  group_by(group) %>% 
  dplyr::count()
```


## Transform variables and re-order levels
```{r}
latency_clean <- latency_clean %>% 
  mutate(participant = factor(participant), 
         object = factor(object), 
         trial_num = factor(trial_num), 
         motion = factor(motion), 
         group = factor(group),
         block = factor(block), 
         cueCond = factor(cueCond), 
         video_phase = factor(video_phase), 
         language = factor(language), 
         value = as.numeric(value)) %>% 
  mutate(motion = fct_relevel(motion, "static", "function"),
         group = fct_relevel(group, "TYP", "ASD"), 
         cueCond = fct_relevel(cueCond, "eye", "arw"))
```



## Visual analysis: distribution of DV
```{r}
ggplot(latency_clean, aes(x = value, fill = cueCond)) + 
  geom_density(alpha = .3) + 
  facet_wrap(~ motion + video_phase) 
```



## Mixed models: full group of 41 TD and 22 ASD
```{r}
# comparing models with and without slope
m_latency = lmerTest::lmer(log(value) ~ cueCond * motion * group + 
                        (1 | participant),
                        data = latency_clean, REML = F) 

m_latency_withslope = lmerTest::lmer(log(value) ~ cueCond * motion * group + 
                        (1 + cueCond | participant),
                        data = latency_clean, REML = F) 

anova(m_latency, m_latency_withslope)



# final model for p values calculated with Kenward-Roger's method
m_latency_REML = lmerTest::lmer(log(value) ~ cueCond * motion * group + 
                        (1 + cueCond | participant),
                        data = latency_clean) 

anova(m_latency_REML, type = c("III"), ddf = "Kenward-Roger")



# confidence interval of cue comparison: estimate +/- 1.96*SE
object1 <- emmeans(m_latency_REML, ~ cueCond)
pairs(object1)

# gaze - arrow
-0.05155349 - (1.96*(0.06751998))
-0.05155349 + (1.96*(0.06751998))


# Cohen's d for cue comparison of gaze-arrow
summarySEwithin(latency_clean, measurevar = "value", 
                withinvars = c("cueCond"),
                idvar = "participant", na.rm = T, conf.interval=.95)

n <- c(208, 203)
mean <- c(0.9320721, 0.9487586)
sd <- c(0.7963495, 0.7214768)
df <- data.frame(n,mean,sd)
pooled.sd1 <- sqrt(sum(df$sd^2 * (df$n - 1)) / (sum(df$n - 1)) )
mes2(mean[1], mean[2], pooled.sd1, n[1], n[2])



# MODEL DIAGNOSTICS
plot_model(m_latency_REML, type = "diag")
```


## Visual analysis: language, block to look at cue condition effects
```{r}
# language 
ggplot(latency_clean, aes(group, value, fill = cueCond)) + 
  geom_boxplot() + 
  geom_jitter(alpha = .3) + 
  stat_summary(fun.data = give.n, geom = "text", size = 6, position=position_dodge(width=.9)) +
  stat_summary(fun.y = mean, shape = 16, col='red', geom = 'point', size = 3, position=position_dodge(width=.9)) + 
  facet_wrap(~ language) 

# block
ggplot(latency_clean, aes(group, value, fill = cueCond)) + 
  geom_boxplot() + 
  geom_jitter(alpha = .3) + 
  stat_summary(fun.data = give.n, geom = "text", size = 6, position=position_dodge(width=.9)) +
  stat_summary(fun.y = mean, shape = 16, col='red', geom = 'point', size = 3, position=position_dodge(width=.9)) + 
  facet_wrap(~ block + video_phase, ncol = 2) 
```
